split classes into library
* split off the class definition and add additional
  class definitions to allow for more modular approach.
* Also looking to add an easy way to switch the backend
  data storage into a database.

Related-To: https://github.com/redhat-performance/quads/issues/25
